requirements.txt
discord.py
aiohttp
python-dotenv
pyfiglet
colorama
pynacl